backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.1
pydantic==1.10.14
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0
google-cloud-workflows>=1.11.0
google-cloud-aiplatform>=1.51.0
